refactor: move backup broker api to break cycle
The scheduler requires the usage of BackupRequestHandler to avoid duplicating all of the requests. Moved ownership to `zeebe-backup` module to make this reusable. Otherwise, there is a cyclic dependency between the gateway and the backups
