+++ zeebe/backup/pom.xml
@@ -96,8 +96,13 @@
     </dependency>
 
     <dependency>
+      <groupId>org.agrona</groupId>
+      <artifactId>agrona</artifactId>
+    </dependency>
+
+    <dependency>
       <groupId>io.camunda</groupId>
-      <artifactId>zeebe-gateway-grpc</artifactId>
+      <artifactId>zeebe-transport</artifactId>
     </dependency>
 
     <dependency>
