+++ zeebe/gateway-grpc/pom.xml
@@ -337,6 +337,12 @@
       <artifactId>camunda-service</artifactId>
     </dependency>
 
+    <dependency>
+      <groupId>io.camunda</groupId>
+      <artifactId>zeebe-backup</artifactId>
+      <scope>test</scope>
+    </dependency>
+
   </dependencies>
 
   <build>
