+++ zeebe/gateway-grpc/src/test/java/io/camunda/zeebe/gateway/admin/backup/BackupQueryStub.java
@@ -7,6 +7,7 @@
  */
 package io.camunda.zeebe.gateway.admin.backup;
 
+import io.camunda.zeebe.backup.client.api.BackupStatusRequest;
 import io.camunda.zeebe.broker.client.api.dto.BrokerError;
 import io.camunda.zeebe.broker.client.api.dto.BrokerErrorResponse;
 import io.camunda.zeebe.broker.client.api.dto.BrokerResponse;
